Adds new landing images and updates components
Introduces new webp images for optimized loading of landing pages,
with different versions for various languages. Updates the image source
in MobileSchool component to use the new assets.

Enhances OnlyNet page with a new conversion section, including
responsive styles and content for promoting Internet, TV, and Mobile
bundles with a call-to-action button and new images for desktop
and mobile views.